+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade `cvxpy` dependency to 1.8.1 or up, and `cvxopt` to 1.3.3 or up to
   ensure Python 3.14 compatibility. Note that individual solver backends may
   still have narrower requirements.
+- Add Knitro (commercial solver) and cuOpt (GPU-accelerated solver) to backends.
+  ([#140](https://github.com/mjpieters/rummikub-solver/issues/140),
+   [#141](https://github.com/mjpieters/rummikub-solver/issues/141)).
 
 ## [1.0.0] - 2025-06-28